pyproject.toml
[project]
name = "bizintel"
version = "0.1.0"
description = "AI-powered Startup Intelligence Engine using RAG"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
	"chromadb>=1.5.5",
	"faiss-cpu>=1.13.2",
	"matplotlib>=3.10.8",
	"openai>=2.26.0",
	"openpyxl>=3.1.5",
	"pandas>=2.2.0",
	"pydantic>=2.12.5",
	"python-dotenv>=1.2.2",
	"sentence-transformers>=5.3.0",
	"streamlit>=1.55.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/bizintel"]
